Changed parameter list of SetAnalyzer to new API from WDSP 1.20

+++ receiver.c
@@ -806,8 +806,8 @@ static void init_analyzer(RECEIVER *rx) {
     double kaiser_pi = 14.0;
     int overlap = 2048;
     int clip = 0;
-    int span_clip_l = 0;
-    int span_clip_h = 0;
+    double span_clip_l = 0;
+    double span_clip_h = 0;
     int pixels=rx->pixels;
     int stitches = 1;
     int calibration_data_set = 0;
